git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@18686 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/tiff.c
@@ -981,7 +981,7 @@ static void TIFFReadPhotoshopLayers(Image* image,const ImageInfo *image_info,
       break;
   }
   i+=4;
-  if (i >= layer_info->length-8)
+  if (i >= (ssize_t) (layer_info->length-8))
     return;
   layers=CloneImage(image,image->columns,image->rows,MagickTrue,exception);
   RemoveImageProfile(layers,"tiff:37724");
